pyproject.toml
[project]
name = "hope-country-workspace"
version = "0.1.0"
description = "HOPE Country Workspace (HCW)"
authors = [
    {name = "sax", email = "s.apostolico@gmail.com"},
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "celery>=5.4.0",
    "deepdiff>=8.0.1",
    "deprecation>=2.1.0",
    "dictdiffer>=0.9.0",
    "django-adminactions>=2.3.0",
    "django-adminfilters>=2.5.0",
    "django-cacheops>=7.1",
    "django-celery-beat>=2.6.0",
    "django-celery-results>=2.5.1",
    "django-concurrency>=2.6.0",
    "django-constance>=3.1.0",
    "django-csp",
    "django-debug-toolbar>=4.4.2",
    "django-environ>=0.11.2",
    "django-flags>=5.0.13",
    "django-jsoneditor>=0.2.4",
    "django-mptt>=0.16.0",
    "django-regex>=0.5.0",
    "django-reversion>=5.1.0",
    "django-select2",
    "django-smart-admin>=2.6.0",
    "django-smart-env>=0.1.0",
    "django-storages[azure]>=1.14.4",
    "django-stubs-ext",
    "django-tailwind>=3.8.0",
    "django>=5.1",
    "djangorestframework>=3.15.1",
    "flower>=2.0.1",
    "hope-flex-fields>=0.6.2",
    "hope-smart-export>=0.3.0",
    "hope-smart-import>=0.3.0",
    "openpyxl>=3.1.5",
    "psycopg2-binary>=2.9.9",
    "python-redis-lock[django]>=4.0.0",
    "redis",
    "sentry-sdk>=2.7.1",
    "social-auth-app-django",
    "unicef-security>=1.5.1",
]
[project.optional-dependencies]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.5.36",
    "mkdocs-awesome-pages-plugin>=2.9.3",
    "mkdocstrings-python",
    "mkdocs-gen-files>=0.5.0",
]

distribution = [
    "uwsgi>=2.0.27",
]

[tool.uv]
package = true
dev-dependencies = [
    "black>=24.4.2",
    "bump2version>=1.0.1",
    "django-browser-reload>=1.17.0",
    "django-webtest>=1.9.11",
    "factory-boy>=3.3.0",
    "flake8>=7.1.0",
    "freezegun>=1.5.1",
    "ipython>=8.29.0",
    "isort>=5.13.2",
    "mypy>=1.11.2",
    "pdbpp>=0.10.3",
    "pip>=24.3.1",
    "pre-commit>=3.7.1",
    "pyquery>=2.0.1",
    "pytest-celery>=1.0.0",
    "pytest-coverage>=0.0",
    "pytest-django>=4.8.0",
    "pytest-echo>=1.7.3",
    "pytest-factoryboy>=2.7.0",
    "pytest-selenium>=4.1.0",
    "pytest-xdist>=3.6.1",
    "pytest>=8.2.2",
    "responses>=0.25.3",
    "types-python-dateutil>=2.9.0.20241003",
    "types-requests>=2.31.0.6",
    "vcrpy>=6.0.2",
]
[tool.uv.sources]

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.pytest_cache
  | \.tox
  | \.venv
  | ~build
  | build
  | ops
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 120
default_section = "THIRDPARTY"
known_first_party = []
known_django = "django"
sections = ["FUTURE","STDLIB","DJANGO","THIRDPARTY","FIRSTPARTY","LOCALFOLDER"]
include_trailing_comma = true
skip = ["migrations", "snapshots", ".venv"]


[tool.django-stubs]
django_settings_module = "country_workspace.config.settings"

#[build-system]
#requires = ["setuptools"]
#build-backend = "setuptools.build_meta"
#
#[build-system]
#requires = ["hatchling"]
#build-backend = "hatchling.build"
#
#[tool.hatch.build.targets.sdist]
#include = [
#  "/src",
#]
#
#[tool.hatch.build.targets.wheel]
#packages = ["src/"]
#[tool.hatch.build]
#include = [
#  "README.md",
#  "src/**/*.csv",
#  "src/**/*.html",
#  "src/**/*.js",
#  "src/**/*.py",
#]
#exclude = [
#  "tests/**",
#]
